requirements_udf.txt
snowflake-snowpark-python>=1.15.0
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99 